+++ libs/feature/audit-builder/src/lib/schema/step-property.constants.ts
@@ -1,0 +1,42 @@
+export const INPUT_TYPE = {
+  STRING: 'string',
+  NUMBER: 'number',
+  BOOLEAN: 'boolean',
+  OPTIONS: 'options',
+  STRING_ARRAY: 'stringArray',
+  RECORDS: 'records'
+} as const;
+
+export const PROPERTY_NAME = {
+  TYPE: 'type',
+  TIMEOUT: 'timeout',
+  VALUE: 'value',
+  SELECTORS: 'selectors',
+  ATTRIBUTES: 'attributes',
+  COUNT: 'count',
+  VISIBLE: 'visible',
+  OPERATOR: 'operator',
+  ASSERTED_EVENTS: 'assertedEvents',
+  FRAME: 'frame',
+  EXPRESSION: 'expression',
+  TARGET: 'target',
+  PROPERTIES: 'properties',
+  BUTTON: 'button',
+  DEVICE_TYPE: 'deviceType',
+  DURATION: 'duration',
+  OFFSET_X: 'offsetX',
+  OFFSET_Y: 'offsetY',
+  DOWNLOAD: 'download',
+  LATENCY: 'latency',
+  UPLOAD: 'upload',
+  KEY: 'key',
+  URL: 'url',
+  X: 'x',
+  Y: 'y',
+  DEVICE_SCALE_FACTOR: 'deviceScaleFactor',
+  HAS_TOUCH: 'hasTouch',
+  HEIGHT: 'height',
+  IS_LANDSCAPE: 'isLandscape',
+  IS_MOBILE: 'isMobile',
+  WIDTH: 'width',
+} as const;
